Moved decoders to a different folder so that all the level related classes will be placed in only one file.
